Add used system memory statistics to the test.
The formula used (hblkhd + arena) is for the system memory, i.e., how
many bytes were allocated from the system (thus it should be multiple of
the page size).  I didn't manage to account for used memory, i.e., how
much of that was actually used.  It seems it should be the difference of
(hblkhd + uordblks - fordblks - keepcost) before and after, but this
yields negative result sometimes, and doesn't match what's printed by
malloc_stats() (system memory does).  Nevertheless, internal fragmentation
seems to be small, so system memory is a good estimate.